requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
procrastinate>=3.5.0
psycopg[binary]>=3.2.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
httpx>=0.26.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger-thrift>=1.21.0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
deprecated>=1.2.14
